pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pandas>=2.2.3",
    "flask-wtf>=1.2.2",
    "trafilatura>=2.0.0",
    "openai>=1.76.2",
    "requests>=2.32.3",
    "python-docx>=1.1.2",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "flask-limiter>=3.12",
    "docx>=0.2.4",
    "python-magic",
    "pyjwt>=2.10.1",
    "pydantic>=2.11.4",
    "flask-dance>=7.1.0",
    "oauthlib>=3.2.2",
    "flask-jwt-extended>=4.7.1",
]
